refactor(extends): outsource to @dargmuesli/nuxt-vio
