--- composables/useHost.ts
@@ -1,9 +1,0 @@
-export const useHost = () => {
-  if (process.server) {
-    const event = useRequestEvent()
-
-    return getHost(event.node.req)
-  } else {
-    return location.host
-  }
-}
