--- server/api/healthcheck.ts
@@ -1,7 +1,0 @@
-import { defineEventHandler } from 'h3'
-
-export default defineEventHandler((event) => {
-  const { res } = event.node
-  res.setHeader('Content-Type', 'text/plain')
-  res.end('OK')
-})
